src/requirements.txt
beautifulsoup4
fastavro
hdfs
kafka-python
pymongo
python-dotenv
requests

websocket-client
git+https://github.com/LJaremek/xtb.git
